Deleted vendor/composer/autoload_classmap.php,...

See merge request hananmohammed2468/bot!10

+++ app/Http/Controllers/Dashboard/DepartmentController.php
@@ -20,6 +20,9 @@ class DepartmentController extends Controller
      */
     public function index(Request $request)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'read departments')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         $branches = Department::departmentData($request);
         return DepartmentsResource::collection($branches);
     }
@@ -42,6 +45,9 @@ class DepartmentController extends Controller
      */
     public function store(DepartmentRequest $request)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'write departments')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         try{
             Department::create($request->all());
             return $this->returnSuccessMessage('Department Stored Succeffully');
@@ -69,6 +75,9 @@ class DepartmentController extends Controller
      */
     public function edit($id)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'update departments')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         try{
             $departments=Department::findOrFail($id);
             if($departments){
@@ -88,6 +97,9 @@ class DepartmentController extends Controller
      */
     public function update(Department $department,DepartmentRequest $request)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'update departments')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         try{
             if($department){
                $department->update($request->all());
@@ -106,6 +118,9 @@ class DepartmentController extends Controller
      */
     public function destroy(Department $department)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'delete departments')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         try{
             if($department){
                 $department->delete();
